Modify(sungje/Layout-boardPage):show-6-post-in-board-page #30
