+++ moae-time/src/apis/dummyBoard.json
@@ -25,7 +25,7 @@
       "views": "488회"
     },
     {
-      "postId": 1,
+      "postId": 4,
       "postTitle": "Hi Front End",
       "writer": "한결",
       "dateAndTime": "02/01 14:32",
@@ -33,7 +33,39 @@
       "views": "587회"
     },
     {
-      "postId": 2,
+      "postId": 5,
+      "postTitle": "Hi Back End",
+      "writer": "승범",
+      "dateAndTime": "01/31 18:52",
+      "conmmentNum": "78개",
+      "views": "389회"
+    },
+    {
+      "postId": 6,
+      "postTitle": "Hi Design",
+      "writer": "재희",
+      "dateAndTime": "02/02 09:00",
+      "conmmentNum": "117개",
+      "views": "488회"
+    },
+    {
+      "postId": 7,
+      "postTitle": "2페이지에 보일 녀석",
+      "writer": "1페이지에 난 안보여야 해",
+      "dateAndTime": "02/02 09:00",
+      "conmmentNum": "117개",
+      "views": "488회"
+    },
+    {
+      "postId": 8,
+      "postTitle": "Hi Front End",
+      "writer": "한결",
+      "dateAndTime": "02/01 14:32",
+      "conmmentNum": "128개",
+      "views": "587회"
+    },
+    {
+      "postId": 9,
       "postTitle": "Hi Back End",
       "writer": "승범",
       "dateAndTime": "01/31 18:52",
@@ -50,7 +82,31 @@
     },
     {
       "postId": 4,
-      "postTitle": "2페이지에 보일 녀석",
+      "postTitle": "Hi Front End",
+      "writer": "한결",
+      "dateAndTime": "02/01 14:32",
+      "conmmentNum": "128개",
+      "views": "587회"
+    },
+    {
+      "postId": 5,
+      "postTitle": "Hi Back End",
+      "writer": "승범",
+      "dateAndTime": "01/31 18:52",
+      "conmmentNum": "78개",
+      "views": "389회"
+    },
+    {
+      "postId": 6,
+      "postTitle": "Hi Design",
+      "writer": "재희",
+      "dateAndTime": "02/02 09:00",
+      "conmmentNum": "117개",
+      "views": "488회"
+    },
+    {
+      "postId": 7,
+      "postTitle": "3페이지에 보일 녀석",
       "writer": "1페이지에 난 안보여야 해",
       "dateAndTime": "02/02 09:00",
       "conmmentNum": "117개",
