+++ moae-time/src/pages/board/BoardPageContainer.jsx
@@ -1,0 +1,15 @@
+import { React, useState, useEffect } from 'react';
+import BoardPagePresenter from './BoardPagePresenter';
+import dummyBoard from '../../apis/dummyBoard.json';
+
+
+function BoardPage() {
+  
+  const [state, setState] = useState()
+
+  return (
+  <BoardPagePresenter />
+  );
+}
+
+export default BoardPage;
