+++ moae-time/src/pages/index.js
@@ -1,5 +1,5 @@
 import HomePage from './home/HomePageContainer';
-import BoardPage from './board/BoardPagePresenter';
+import BoardPage from './board/BoardPageContainer';
 import MyPage from './mypage/MyPageContainer';
 import BoardOne from './boardOne/BoardOne';
 import BoardEditPresenter from './board-edit/BoardEditPresenter';
